+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## 2025-07-23
+- 1621 Added "Use Animated Sprites" option with new sprite models and animations
 - 2214 Break Game.js into smaller initialization modules for easier maintenance
 - 0000 Rewrite mats.json to load building modules from `js/mats`